refactor(foundation-demo): remove unnecessary shipping component
remove unnecessary shipping component
