--- apps/foundation-demo/src/app/checkout/components/shipping/shipping.component.html
@@ -1,13 +1,0 @@
-<div class="shipping" shipping-container #ShippingContainer="ShippingContainer">
-  <shipping-form 
-    [hidden]="!showShippingForm"
-    [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
-    (updateShippingInfo)="ShippingContainer.updateShippingInfo($event); toggleShippingView()"></shipping-form>
-  <shipping-summary 
-    [hidden]="showShippingForm" 
-    *ngIf="ShippingContainer.shippingInfo$ | async" 
-    [shippingInfo]="ShippingContainer.shippingInfo$ | async" 
-    [shippingOption]="ShippingContainer.shippingOption$ | async"
-    (editShippingInfo)="toggleShippingView()"
-    (updateShippingOption)="ShippingContainer.updateShippingOption($event)"></shipping-summary>
-</div>
